requirements.txt
faiss-cpu
packaging
fastapi
uvicorn
torch
PIL
torchvision
lightning
